requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.1
filelock==3.20.0
fonttools==4.62.0
fsspec==2025.12.0
h11==0.16.0
idna==3.11
Jinja2==3.1.3
kiwisolver==1.5.0
lapx==0.9.4
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
opencv-python==4.13.0.92
packaging==26.0
pillow==12.0.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
six==1.17.0
starlette==0.52.1
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
